Material Garage: moving to a broader solution.
The 09/01 update brings early (none to this day) support to the Athmo
Platform Kit, bringing an health solution to the HomeKit Platform. With
Athmo, you'll be able to check not only the temperature of your house
(Athmo can be placed both indoors and outdoors), but also other
important infos like humidity, pressure and even the dew point.

Other changes in the 09012016 build:
- Removed landscape layout. As HomeKit is moving towards a
"portrait-only" approach, we felt it was unnecessary.
- Improved Italian translation.