HomeCloud: Making all the HPKs feel at home.
With the 09 Jan 2016 build of Garage, we announced an addition to the
HPK family: Athmo, the all-in-one solution for home users to check the
temperature, pressure, altitude and so much more of their houses. Or
other places. The app, however, stayed the same. With the 50th build of
Garage/Homekit and version 4.3 of the Material Redesign, We're
announcing a radical change in the app.
Hello, HomeCloud.