HomeCloud v6.0!
The full description of this update will be delivered when it is
complete (with Version 6.0.1). This version is not quite ready for
prime-time yet, so if you're still using version 5.2, continue using it
as it is a well tested version ready to be used on the streets. If
you're curious, though, version 6.0 will bring a new way to unlock your
garage door.